requirements.txt
tqdm==4.20.0
psycopg2-binary
sqlalchemy
IPython
pybind11
numpy>=1.16.0
scipy
pandas
Cython==0.29.2
pybind11
pandas
six
scikit-learn>=0.20.0
sklearn-contrib-lightning
ipykernel
seaborn
